+++ SecretSanta/SecretSanta.sln
@@ -15,11 +15,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Web", "src\Secr
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{EA18F841-87B0-4A2A-87BE-AEED9D0F0164}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Business.Tests", "SecretSanta.Business.Gift.Tests\SecretSanta.Business.Tests.csproj", "{37816553-D99F-4674-93CF-D11B269F5687}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Business.Tests2", "SecretSanta.Business.Gift.Tests\SecretSanta.Business.Tests2.csproj", "{37816553-D99F-4674-93CF-D11B269F5687}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Api.Tests", "test\SecretSanta.Api.Tests\SecretSanta.Api.Tests.csproj", "{48F13FA4-3255-406F-B480-44CFDF48251A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Api.Tests", "test\SecretSanta.Api.Tests\SecretSanta.Api.Tests.csproj", "{48F13FA4-3255-406F-B480-44CFDF48251A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Data.Tests", "test\SecretSanta.Data.Tests\SecretSanta.Data.Tests.csproj", "{32F2C37F-5A2A-4D46-B7D6-7A36A9F38742}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Data.Tests", "test\SecretSanta.Data.Tests\SecretSanta.Data.Tests.csproj", "{32F2C37F-5A2A-4D46-B7D6-7A36A9F38742}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Business.Tests", "test\SecretSanta.Business.Tests\SecretSanta.Business.Tests.csproj", "{AE9C0D62-7CC1-4265-A126-3F9CE8A35A80}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -55,6 +57,10 @@ Global
 		{32F2C37F-5A2A-4D46-B7D6-7A36A9F38742}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{32F2C37F-5A2A-4D46-B7D6-7A36A9F38742}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{32F2C37F-5A2A-4D46-B7D6-7A36A9F38742}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE9C0D62-7CC1-4265-A126-3F9CE8A35A80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE9C0D62-7CC1-4265-A126-3F9CE8A35A80}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE9C0D62-7CC1-4265-A126-3F9CE8A35A80}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE9C0D62-7CC1-4265-A126-3F9CE8A35A80}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -67,6 +73,7 @@ Global
 		{37816553-D99F-4674-93CF-D11B269F5687} = {EA18F841-87B0-4A2A-87BE-AEED9D0F0164}
 		{48F13FA4-3255-406F-B480-44CFDF48251A} = {EA18F841-87B0-4A2A-87BE-AEED9D0F0164}
 		{32F2C37F-5A2A-4D46-B7D6-7A36A9F38742} = {EA18F841-87B0-4A2A-87BE-AEED9D0F0164}
+		{AE9C0D62-7CC1-4265-A126-3F9CE8A35A80} = {EA18F841-87B0-4A2A-87BE-AEED9D0F0164}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E2234040-21B9-4AF6-9A6A-3ABE8D5F86E0}
